Add CMakeLists.txt file, add Qt 6 build support
This adds a CMake project file that supports both Qt 5 and 6 builds.

Qt 6 build uses QML modules, defined from CMake.
Resources aside from qml are still kept in a separate resource file.

QML registration done the old way remains.

NOTE: when built for Qt 6, it has visual artifacts that I have not yet
figured out. The watermelon texture does not show rendered when the app
starts (well, it does briefly, but then disappears). If you move one of
the sliders, you will see that the scene is being rendered, but it has a
square in which it does not render - which happens to cover the entire
square region that the texture is initially drawn in.